:package: improved code
